+++ ky3m/services/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ['mm_storage', 'pickler', 'extractor', 'mcmod', 'jar_keeper']
+__all__ = ['mm_storage', 'pickler', 'extractor', 'mcmod', 'jar_keeper', 'bundler']