+++ CHANGELOG.md
@@ -9,6 +9,7 @@ These are the changes made to the Ky3M project from the beginning:
 - Recast unknown versions processing.
 - Add RELEASE, PUNISH, PEER and ADOPT bad saved ID handling.
 - Update APPLY output.
+- Refactor bundle-methods, some code is moved to the functions of the new module named "bundle.py".
 
 ## [Release 0.2.1](https://github.com/Procybit/Ky3m/releases/tag/v0.2.1) (2022-08-17) - v0.2.0 Important patches